+++ examples/image.py
@@ -32,4 +32,4 @@ def app():
 
 
 if __name__ == "__main__":
-    c.main("Image Viewer", app(), 500, 500)
+    c.main(app(), "Image Viewer")