+++ examples/todo.py
@@ -33,4 +33,4 @@ def app():
 
 
 if __name__ == "__main__":
-    c.main("Todo List", app(), 500, 500)
+    c.main(app(), "Todo List")